custom-elements/registering-custom-elements/definition-construction-algorithm-test-002.html

<!DOCTYPE html>
<html>
<head>
<title>Default namespace is HTML namespace</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Default namespace is HTML namespace">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    var GeneratedConstructor = doc.register('x-a-1');

    var customElement = new GeneratedConstructor();

    assert_equals(customElement.namespaceURI, HTML_NAMESPACE, 'Wrong namespace');

}, 'Default namespace is HTML namespace');


test(function() {
    for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
        var element = HTML5_ELEMENTS[i];
            
        var doc = newHTMLDocument();
        
        var obj = doc.createElement(element);
        
        var name = 'x-' + element;
        
        var proto = Object.create(obj.constructor.prototype);
        
        var GeneratedConstructor = doc.register(name, {
            prototype: proto, 
            extends: element
        });
        
        var customElement = new GeneratedConstructor();

        assert_equals(customElement.namespaceURI, HTML_NAMESPACE, 'Wrong namespace');    
    }

}, 'Default namespace is HTML namespace. Test constructor of extended HTML element');


</script>
</body>
</html>
